fix: refactor validators
